LayoutTests/canvas/philip/tests/size.attributes.parse.percent-expected.txt

< [index] >
size.attributes.parse.percent
Parsing of non-negative integers
References: size.nonnegativeinteger
Actual output:
Expected output:

Passed
